backend/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-pinecone>=0.0.3

# Vector DB
pinecone-client>=3.0.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
beautifulsoup4>=4.12.0
lxml>=5.0.0
readability-lxml>=0.8.1

# Video/Audio processing
ffmpeg-python>=0.2.0
openai-whisper>=20231117
pillow>=10.0.0

# Embeddings
sentence-transformers>=2.3.0
torch>=2.1.0
open-clip-torch>=2.24.0

# Reranking
rank-bm25>=0.2.2

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
aiofiles>=23.2.0
tenacity>=8.2.3

# Monitoring
langsmith>=0.0.87
prometheus-client>=0.19.0

# MCP
mcp>=0.1.0
